+++ core/data/src/main/java/dev/enesky/core/data/AnimeFilter.kt
@@ -14,27 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.enesky.core.network.model
-
-import com.google.gson.annotations.SerializedName
+package dev.enesky.core.data
 
 /**
  * Created by Enes Kamil YILMAZ on 28/10/2023
  */
-data class AnimePagingResponse(
-    val pagination: Pagination,
-    val data: List<Anime>,
-)
 
-data class Pagination(
-    @SerializedName("last_visible_page") val lastVisiblePage: Int,
-    @SerializedName("has_next_page") val hasNextPage: Boolean,
-    @SerializedName("current_page") val currentPage: Int,
-    val items: PaginationItems,
-)
-
-data class PaginationItems(
-    val count: Int,
-    val total: Int,
-    @SerializedName("per_page") val perPage: Int,
-)
+enum class AnimeFilter(val filter: String) {
+    AIRING("airing"),
+    UPCOMING("upcoming"),
+    POPULARITY("bypopularity"),
+    FAVORITE("favorite"),
+}
