+++ core/data/src/main/java/dev/enesky/core/data/AnimeRating.kt
@@ -14,19 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.enesky.core.network.model
+package dev.enesky.core.data
 
 /**
- * Created by Enes Kamil YILMAZ on 28/10/2023
+ * Created by Enes Kamil YILMAZ on 04/12/2023
  */
 
-@Suppress("unused")
-enum class AnimeType(val type: String) {
-
-    TV("tv"),
-    MOVIE("movie"),
-    OVA("ova"),
-    SPECIAL("special"),
-    ONA("ona"),
-    MUSIC("music"),
+enum class AnimeRating(val rating: String) {
+    G("g"), // All ages
+    PG("pg"), // Children
+    PG13("pg13"), // Teens 13 or older
+    R17("r17"), // +17 (Violence & Profanity)
+    R("r"), // Mild Nudity
+    RX("rx"), // Hentai
 }
