+++ core/data/src/main/java/dev/enesky/core/data/FullAnime.kt
@@ -1,4 +1,4 @@
-package dev.enesky.core.network.model
+package dev.enesky.core.data
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
