+++ core/domain/src/main/java/dev/enesky/core/domain/usecase/AnimeCharactersUseCase.kt
@@ -16,7 +16,7 @@
  */
 package dev.enesky.core.domain.usecase
 
-import dev.enesky.core.network.model.Character
+import dev.enesky.core.data.Character
 
 /**
  * Created by Enes Kamil YILMAZ on 29/10/2023
