+++ core/network/src/main/java/dev/enesky/core/network/di/RepositoryModule.kt
@@ -16,7 +16,7 @@
  */
 package dev.enesky.core.network.di
 
-import dev.enesky.core.network.api.service.JikanService
+import dev.enesky.core.network.api.JikanService
 import dev.enesky.core.network.paging.TopAnimePagingSource
 import dev.enesky.core.network.repository.JikanDataSource
 import dev.enesky.core.network.repository.JikanRepository
