+++ core/network/src/main/java/dev/enesky/core/network/paging/TopAnimePagingSource.kt
@@ -19,9 +19,9 @@ package dev.enesky.core.network.paging
 import android.util.Log
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import dev.enesky.core.network.api.service.JikanService
-import dev.enesky.core.network.model.Anime
-import dev.enesky.core.network.model.AnimeFilter
+import dev.enesky.core.network.api.JikanService
+import dev.enesky.core.data.Anime
+import dev.enesky.core.data.AnimeFilter
 import retrofit2.HttpException
 import java.io.IOException
 
@@ -31,14 +31,14 @@ import java.io.IOException
 
 class TopAnimePagingSource(
     private val jikanService: JikanService,
-    private val animeFilter: AnimeFilter = AnimeFilter.POPULARITY,
-) : PagingSource<Int, Anime>() {
+    private val animeFilter: dev.enesky.core.data.AnimeFilter = dev.enesky.core.data.AnimeFilter.POPULARITY,
+) : PagingSource<Int, dev.enesky.core.data.Anime>() {
 
-    override fun getRefreshKey(state: PagingState<Int, Anime>): Int? {
+    override fun getRefreshKey(state: PagingState<Int, dev.enesky.core.data.Anime>): Int? {
         return state.anchorPosition
     }
 
-    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Anime> {
+    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, dev.enesky.core.data.Anime> {
         return try {
             val nextPage = params.key ?: 1
             val animePagingResponse = jikanService.getTopAnimePaging(
