+++ core/network/src/main/java/dev/enesky/core/network/repository/JikanDataSource.kt
@@ -17,11 +17,11 @@
 package dev.enesky.core.network.repository
 
 import androidx.paging.PagingData
-import dev.enesky.core.network.model.Anime
-import dev.enesky.core.network.model.AnimeFilter
-import dev.enesky.core.network.model.BaseResponse
-import dev.enesky.core.network.model.Character
-import dev.enesky.core.network.model.FullAnime
+import dev.enesky.core.data.Anime
+import dev.enesky.core.data.AnimeFilter
+import dev.enesky.core.data.BaseResponse
+import dev.enesky.core.data.Character
+import dev.enesky.core.data.FullAnime
 import kotlinx.coroutines.flow.Flow
 
 /**
@@ -30,7 +30,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface JikanDataSource {
 
-    fun getTopAnimePagingData(animeFilter: AnimeFilter): Flow<PagingData<Anime>>
-    suspend fun getAnimeById(animeId: Int): Result<BaseResponse<FullAnime>>
-    suspend fun getCharactersByAnimeId(animeId: Int): Result<List<Character>>
+    fun getTopAnimePagingData(animeFilter: dev.enesky.core.data.AnimeFilter): Flow<PagingData<dev.enesky.core.data.Anime>>
+    suspend fun getAnimeById(animeId: Int): Result<dev.enesky.core.data.BaseResponse<dev.enesky.core.data.FullAnime>>
+    suspend fun getCharactersByAnimeId(animeId: Int): Result<List<dev.enesky.core.data.Character>>
 }
