+++ feature/home/build.gradle.kts
@@ -26,4 +26,5 @@ android.namespace = "dev.enesky.feature.home"
 dependencies {
     implementation(projects.core.network)
     implementation(projects.core.domain)
+    implementation(projects.core.data)
 }
