+++ feature/home/src/main/java/dev/enesky/feature/home/HomeViewModel.kt
@@ -9,9 +9,9 @@ import dev.enesky.core.common.delegate.UiState
 import dev.enesky.core.common.delegate.UiStateDelegate
 import dev.enesky.core.domain.usecase.AnimeUseCase
 import dev.enesky.core.domain.usecase.TopAnimePagingUseCase
-import dev.enesky.core.network.model.Anime
-import dev.enesky.core.network.model.AnimeFilter
-import dev.enesky.core.network.model.asMiniAnime
+import dev.enesky.core.data.Anime
+import dev.enesky.core.data.AnimeFilter
+import dev.enesky.core.data.asMiniAnime
 import dev.enesky.core.network.util.Resource
 import dev.enesky.core.network.util.asResource
 import dev.enesky.core.ui.mapper.pagingMap
